backend/requirements.txt
# ===========================================
# Argus Backend - Python Dependencies
# ===========================================

# Core Framework
Django>=5.2,<6.0

# API Framework (Django Ninja)
django-ninja>=1.3.0,<2.0
django-ninja-jwt>=5.3.0,<6.0
pydantic[email]>=2.7.0,<3.0

# JSON Serialization (Fast)
orjson>=3.10.0,<4.0

# Authentication
PyJWT>=2.8.0,<3.0

# Image Processing
Pillow>=10.0.0,<12.0

# Environment Configuration
python-dotenv>=1.0.0,<2.0

# CORS Headers
django-cors-headers>=4.3.0,<5.0

# WebSocket / Channels
channels>=4.0.0,<5.0
daphne>=4.0.0,<5.0
channels-redis>=4.2.0,<5.0

# AI Integration (Google Gemini)
google-generativeai>=0.5.0,<1.0

# Retry Logic
tenacity>=8.2.0,<10.0

# ===========================================
# Optional Dependencies
# ===========================================

# Redis Cache
django-redis>=5.4.0,<6.0

# ===========================================
# Development Tools (optional)
# ===========================================

# flake8>=7.0.0
# pytest>=8.0.0
# pytest-django>=4.8.0
